requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
crewai==0.30.11
langchain==0.1.20
langchain-google-genai==1.0.3
python-dotenv==1.0.1
pandas==2.1.4
numpy==1.26.4
openpyxl==3.1.2
PyPDF2==3.0.1